refactor: rename standard-form folder to standardForm
