+++ src/components/index.tsx
@@ -3,4 +3,4 @@ export { HidePreventUnmount } from './HidePreventUnmount'
 export * from './metadataSelects'
 export * from './SearchableSingleSelect'
 export * from './sectionList'
-export * from './standard-form'
+export * from './standardForm'
