+++ src/components/standardForm/StandardFormSectionDescription.tsx
@@ -1,6 +1,10 @@
 import React from 'react'
 import classes from './StandardFormSectionDescription.module.css'
 
-export function StandardFormSectionDescription({ children }: { children: React.ReactNode }) {
+export function StandardFormSectionDescription({
+    children,
+}: {
+    children: React.ReactNode
+}) {
     return <p className={classes.standardFormSectionDescription}>{children}</p>
 }
