+++ src/components/standardForm/StandardFormSectionTitle.tsx
@@ -1,6 +1,10 @@
 import React from 'react'
 import classes from './StandardFormSectionTitle.module.css'
 
-export function StandardFormSectionTitle({ children }: { children: React.ReactNode }) {
+export function StandardFormSectionTitle({
+    children,
+}: {
+    children: React.ReactNode
+}) {
     return <h2 className={classes.standardFormSectionTitle}>{children}</h2>
 }
